Disable warning dialog for Wine arch check
Winetricks checks the `wine` and `wineserver` executables for the file
architecture while expecting either a 32 or 64-bit binary. This is not
true on Protontricks as it instead wraps each underlying binary with a
script.

Since this will cause two warning dialogs to pop up every time
Protontricks is used, make Winetricks only print the warning to stderr
instead.